Add Trick detail page + add test
