+++ templates/layout/_footer.html.twig
@@ -1,4 +1,4 @@
-<div class="w-full bg-gray-500 text-white">
+<div class="w-full relative z-10 bg-gray-500 text-white">
   <div class="container max-w-screen-xl mx-auto px-8 md:px-0 py-10 flex flex-col md:flex-row justify-center">
     <div class="w-full text-center">
       <h3 class="pb-3 uppercase">Menu</h3>
