+++ templates/layout/tricks.html.twig
@@ -21,7 +21,7 @@
 {%- endblock -%}
 
 {%- block title -%}
-  Tous les tricks
+  Tous les tricks - SnowTricks
 {%- endblock -%}
 
 {% block body %}
