+++ tests/Controller/IndexControllerTest.php
@@ -13,12 +13,4 @@ class IndexControllerTest extends WebTestCase
 
         self::assertResponseIsSuccessful();
     }
-
-    public function testTricksPage(): void
-    {
-        $client = static::createClient();
-        $client->request('GET', '/');
-
-        self::assertResponseIsSuccessful();
-    }
 }
