+++ tests/Controller/TricksControllerTest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Tests\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class TricksControllerTest extends WebTestCase
+{
+    public function testTricksPage(): void
+    {
+        $client = static::createClient();
+        $client->request('GET', '/');
+
+        self::assertResponseIsSuccessful();
+    }
+
+    public function testTrickDetailPage(): void
+    {
+        $client = static::createClient();
+        $client->request('GET', '/');
+
+        self::assertResponseIsSuccessful();
+    }
+}
